refactor: 合并定时任务入口，日历视图内融合列表检索能力
将设置页的定时任务 tab 合并到侧边栏日历入口 (CronCalendarView)，
统一为一个入口点。日历视图顶部新增"日历/列表"Tab 切换：日历模式保留
月历网格 + 选中日期面板，列表模式复刻原 CronPanel 的搜索框、状态过滤
和行级操作 (立即运行/编辑/启停/删除)。两种模式共享同一份任务数据，
任何操作后通过 refreshAll 同步刷新日历事件和任务列表，保持跨视图一致。

- 删除 src/components/settings/CronPanel.tsx
- 从 SettingsView 的 SECTIONS / SettingsSection 类型中移除 cron
- 清理 settings.cron i18n 键，新增 cron.viewCalendar / cron.viewList
