+++ src/i18n/locales/en.json
@@ -408,7 +408,6 @@
     "memoryInherited": "inherited",
     "memoryOverridden": "overridden",
     "memoryResetToGlobal": "Reset to inherit global settings",
-    "cron": "Scheduled Tasks",
     "notifications": "Notifications",
     "shortcuts": "Shortcuts",
     "sandbox": "Docker Sandbox",
@@ -1245,6 +1244,8 @@
     "noTasksThisDay": "No tasks on this day",
     "tasks": "tasks",
     "today": "Today",
+    "viewCalendar": "Calendar",
+    "viewList": "List",
     "searchPlaceholder": "Search tasks...",
     "filterAll": "All",
     "jobNotFound": "Task not found",
