+++ src/i18n/locales/zh.json
@@ -408,7 +408,6 @@
     "memoryInherited": "继承全局",
     "memoryOverridden": "已覆盖",
     "memoryResetToGlobal": "恢复继承全局设置",
-    "cron": "定时任务",
     "notifications": "通知",
     "shortcuts": "快捷键",
     "sandbox": "Docker 沙箱",
@@ -1245,6 +1244,8 @@
     "noTasksThisDay": "当天没有任务",
     "tasks": "个任务",
     "today": "今天",
+    "viewCalendar": "日历",
+    "viewList": "列表",
     "searchPlaceholder": "搜索任务...",
     "filterAll": "全部",
     "jobNotFound": "任务未找到",
